requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
pydantic==2.5.2
python-dotenv==1.0.0
structlog==23.2.0
